Start splitting off sugar build functions
