+++ src/Exec_PrepareForLeap.h
@@ -1,7 +1,6 @@
 #ifndef INC_EXEC_PREPAREFORLEAP_H
 #define INC_EXEC_PREPAREFORLEAP_H
 #include "Exec.h"
-#include "AtomMap.h"
 #include <set>
 #include <map>
 #include <vector>
@@ -31,10 +30,6 @@ class Exec_PrepareForLeap : public Exec {
 
     
     
-    /// Set a reduced PDB res to glycam map when dat file not found.
-    void SetGlycamPdbResMap();
-    /// Load PDB res to glycam map from dat file
-    int LoadGlycamPdbResMap(std::string const&);
     /// Set PDB residue names recognized by Amber FFs
     void SetPdbResNames();
     /// Load PDB residue names recognized by Amber FFs from dat file
@@ -42,45 +37,9 @@ class Exec_PrepareForLeap : public Exec {
 
     void LeapBond(int,int,Topology const&, CpptrajFile*) const;
 //    int CalcStereocenterTorsion(double&, int, Topology const&, Frame const&) const;
-    int FindRemainingChainCarbons(Iarray&, int, Topology const&, int, Iarray const&) const;
     /// Try to find any missing bonds to C1 atoms
 //    int FindSugarC1Linkages(Sugar const&, Topology&, Frame const&) const;
-    /// Determine orientation around anomeric carbon
-    int CalcAnomericTorsion(double&, int, int, int, Iarray const&,
-                            Topology const&, Frame const&) const;
-    /// Determine orientation around anomeric reference carbon
-    int CalcAnomericRefTorsion(double&, int, int, int, Iarray const&,
-                               Topology const&, Frame const&) const;
-    /// Determine orientation around configurational carbon
-    int CalcConfigCarbonTorsion(double&, int, Iarray const&,
-                                Topology const&, Frame const&) const;
 
-    /// \return Sugar with atom indices set up
-    Sugar IdSugarRing(int, Topology const&) const;
-    /// Change PDB atom names to Glycam names
-    int ChangePdbAtomNamesToGlycam(std::string const&, Residue const&,
-                                   Topology&, FormTypeEnum) const;
-    /// Determine form/chirality for furanose
-    int DetermineUpOrDown(SugarToken&, Sugar const&, Topology const&, Frame const&) const;
-    /// Determine form/chirliaty for pyranose 
-    int DetermineAnomericForm(SugarToken&, Sugar&, Topology const&, Frame const&) const;
-    /// \return Glycam linkage code for given link atoms
-    std::string GlycamLinkageCode(std::set<Link> const&, Topology const&) const;
-    /// Determine linkages for the sugar
-    std::string DetermineSugarLinkages(Sugar const&, CharMask const&, Topology&, ResStatArray&,
-                                       CpptrajFile*, std::set<BondType>&) const;
-    /// Create a residue mask string for selecting Glycam-named sugar residues.
-    std::string GenGlycamResMaskString() const;
-    /// Try to identify sugar name, form, and linkages
-    int IdentifySugar(Sugar&, Topology&, Frame const&, CharMask const&, CpptrajFile*, std::set<BondType>&);
-    /// Try to find missing linkages to anomeric carbon in sugar.
-    int FindSugarC1Linkages(int, int, Topology&, Frame const&) const;
-    
-    /// Attempt to fix any issues with sugars
-    int FixSugarsStructure(std::vector<Sugar>&, std::string const&, Topology&, Frame&,
-                           bool, bool) const;
-
-    int PrepareSugars(std::string const&, std::vector<Sugar>&, Topology&, Frame const&, CpptrajFile*);
     int FindTerByBonds(Topology&, CharMask const&) const;
     int SearchForDisulfides(double, std::string const&, std::string const&, bool,
                             Topology&, Frame const&, CpptrajFile*);
@@ -103,9 +62,6 @@ class Exec_PrepareForLeap : public Exec {
     /// Print a warning for residues that will need modification after leap
     static void LeapFxnGroupWarning(Topology const&, int);
 
-    typedef std::pair<NameType, SugarToken> PairType;
-    typedef std::map<NameType, SugarToken> MapType;
-    MapType pdb_to_glycam_; ///< Map PDB residue names to sugar information tokens
 
     typedef std::set<NameType> SetType;
     SetType pdb_res_names_; ///< PDB residue names recognized by Amber FFs
@@ -134,11 +90,7 @@ class Exec_PrepareForLeap : public Exec {
 
     std::string leapunitname_;
     bool errorsAreFatal_;   ///< If false, try to skip errors.
-    bool hasGlycam_;        ///< If true, assume sugars already have glycam names
-    bool useSugarName_;     ///< If true, base form/chirality on name instead of geometry
     int debug_;             ///< Debug level
     std::string solventResName_; ///< Solvent residue name
-    AtomMap myMap_;
-    
 };
 #endif
